+++ anonymizer.py
@@ -610,16 +610,11 @@ def main():
     install_dir = os.path.dirname(os.path.realpath(__file__))
     run_as_exe = getattr(sys, "frozen", False) and hasattr(sys, "_MEIPASS")
     # TODO: enhance cmd line processing using Click library
-    # TODO: run time log level change
-    # may need user settable log level for pynetdicom and pydicom separately
-    debug_mode = "debug" in args or "DEBUG" in args
-    init_logging(install_dir, debug_mode, run_as_exe)
+    init_logging(install_dir, run_as_exe)
     os.chdir(install_dir)
 
     logger = logging.getLogger()  # get root logger
     logger.info(f"cmd line args={args}")
-    if debug_mode:
-        logger.info(f"DEBUG Mode")
     if run_as_exe:
         logger.info(f"Running as executable (PyInstaller)")
     logger.info(f"Python Optimization Level: {sys.flags.optimize}")